+++ .gitignore
@@ -2,3 +2,5 @@
 /.idea
 *.vtt
 *.srt
+/build/
+/vtt_to_srt.egg-info
